@@ -1,13 +1,14 @@
 {
   "name": "react-datetime",
-  "version": "2.10.2",
+  "version": "3.0.0",
   "description": "A lightweight but complete datetime picker React.js component.",
   "homepage": "https://github.com/YouCanBookMe/react-datetime",
   "repository": {
     "type": "git",
     "url": "https://github.com/YouCanBookMe/react-datetime"
   },
-  "main": "./DateTime.js",
+  "main": "./lib/DateTime.js",
+  "module": "./src/DateTime.js",
   "typings": "./DateTime.d.ts",
   "files": [
     "DateTime.js",
@@ -17,16 +18,19 @@
     "dist"
   ],
   "scripts": {
-    "build:mac": "./node_modules/.bin/gulp",
-    "build:win": "./node_modules/.bin/gulp.cmd",
-    "dev": "./node_modules/.bin/webpack-dev-server --config example/webpack.config.js --devtool eval --progress --colors --hot --content-base example",
-    "lint": "./node_modules/.bin/eslint src/ DateTime.js test/",
-    "test": "./node_modules/.bin/jest",
-    "test:typings": "./node_modules/.bin/tsc -p ./typings",
-    "test:snapshot": "./node_modules/.bin/jest snapshot",
-    "test:snapshot:update": "./node_modules/.bin/jest snapshot --updateSnapshot",
-    "test:all": "npm run test:typings && npm run test",
-    "test:watch": "./node_modules/.bin/jest --watch"
+    "build": "npm run build:bundle && npm run build:lib",
+    "build:bundle": "webpack -p",
+    "build:lib": "rm -rf lib && babel src --out-dir lib --babelrc",
+    "build:lib:watch": "npm run build:lib -- --watch",
+    "dev": "webpack-dev-server --config example/webpack.config.js --devtool eval --progress --colors --hot --content-base example",
+    "lint": "eslint src/ test/",
+    "lint:fix": "npm run lint -- --fix",
+    "test": "jest test/tests.spec.js",
+    "test:watch": "npm test --watch",
+    "test:typings": "tsc -p ./typings",
+    "test:snapshot": "jest test/snapshots.spec.js",
+    "test:snapshot:update": "jest test/snapshots.spec.js --updateSnapshot",
+    "test:all": "npm run test:typings && npm run test"
   },
   "keywords": [
     "react",
@@ -40,47 +44,39 @@
   "license": "MIT",
   "peerDependencies": {
     "moment": ">=2.16.0",
-    "react": ">=0.13",
-    "react-dom": ">=0.13"
+    "react": "^15.0.0",
+    "react-dom": "^15.0.0"
   },
   "devDependencies": {
     "@types/react": ">=15",
-    "babel-core": "^6.22.1",
-    "babel-jest": "^18.0.0",
-    "babel-loader": "^6.2.1",
+    "babel-cli": "^6.22.0",
+    "babel-core": "^6.22.0",
+    "babel-eslint": "^8.0.1",
+    "babel-jest": "^21.2.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-remove-strict-mode": "0.0.2",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.22.0",
-    "enzyme": "^2.7.1",
+    "create-react-class": "^15.6.2",
+    "enzyme": "^2.9.1",
     "eslint": "^3.1.0",
-    "gulp": "^3.9.0",
-    "gulp-babel": "^6.1",
-    "gulp-insert": "^0.4.0",
-    "gulp-plumber": "^1.1.0",
-    "gulp-rename": "^1.2.2",
-    "gulp-sourcemaps": "^2.4.0",
-    "gulp-uglify": "^1.2.0",
-    "jest": "^18.1.0",
-    "jest-cli": "^18.1.0",
-    "jsdom": "^7.0.2",
-    "mocha": "^3.2.0",
+    "eslint-plugin-react": "^7.4.0",
+    "jest": "^21.2.1",
     "moment": ">=2.16.0",
     "pre-commit": "^1.1.3",
-    "react": ">=0.13",
-    "react-addons-test-utils": ">=0.13",
-    "react-dom": ">=0.13",
-    "react-test-renderer": "^15.4.2",
-    "through2": "^2.0.3",
+    "react": "~15.4.0",
+    "react-addons-test-utils": "~15.4.0",
+    "react-dom": "~15.4.0",
+    "react-onclickoutside": "^5.11.1",
+    "react-test-renderer": "~15.4.0",
     "typescript": "^2.0.10",
     "webpack": "^2.2.1",
-    "webpack-dev-server": "^1.7.0",
-    "webpack-stream": "^3.2.0"
+    "webpack-dev-server": "^2.9.1"
   },
   "dependencies": {
-    "create-react-class": "^15.5.2",
-    "object-assign": "^3.0.0",
-    "prop-types": "^15.5.7",
-    "react-onclickoutside": "^5.9.0"
+    "classnames": "^2.2.5",
+    "prop-types": "^15.5.7"
   },
   "pre-commit": [
     "lint",
