@@ -1,14 +1,13 @@
-var DateTime = require('../DateTime.js');
-var React = require('react');
-var ReactDOM = require('react-dom');
+import DateTime from '../src/DateTime'
+import React from 'react'
+import ReactDOM from 'react-dom'
 
-ReactDOM.render(
-  React.createElement(DateTime, {
-    viewMode: 'months',
-    dateFormat: 'MMMM',
-    isValidDate: function(current) {
-      return current.isBefore(DateTime.moment().startOf('month'));
-    }
-  }),
-  document.getElementById('datetime')
-);
+const max = DateTime.moment().startOf('month')
+
+const options = {
+  dateFormat: 'DD/MM/YYYY',
+  timeFormat: 'h:mm:ss SSS A',
+  isValidDate: current => current.isBefore(max)
+}
+
+ReactDOM.render(<DateTime {...options} />, document.getElementById('datetime'))
